+++ implementations/elixir/build.gradle
@@ -3,7 +3,6 @@ def appsDir = 'ockam'
 
 def apps = [
   'ockam',
-  'ockam_transport_udp',
   'ockam_vault_software',
   'ockam_node_web',
   'ockam_hub'
