--- implementations/elixir/ockam/ockam_transport_udp/.formatter.exs
@@ -1,4 +1,0 @@
-# Used by "mix format"
-[
-  inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"]
-]
