--- implementations/elixir/ockam/ockam_transport_udp/.gitignore
@@ -1,2 +1,0 @@
-_build
-deps
