--- implementations/elixir/ockam/ockam_transport_udp/lib/udp.ex
@@ -1,33 +1,0 @@
-defmodule Ockam.Transport.UDP do
-  @moduledoc """
-  Implements the Ockam UDP Transport.
-  """
-
-  use Application
-
-  # Called when the Ockam application is started.
-  #
-  # This function is called when an application is started using
-  # `Application.start/2`, `Application.ensure_started/2` etc.
-  #
-  @doc false
-  def start(_type, _args) do
-    # Specifications of child processes that will be started and supervised.
-    #
-    # See the "Child specification" section in the `Supervisor` module for more
-    # detailed information.
-    children = [
-      Ockam.Transport.UDP.DynamicSupervisor
-    ]
-
-    # Start a supervisor with the given children. The supervisor will inturn
-    # start the given children.
-    #
-    # The :one_for_one supervision strategy is used, if a child process
-    # terminates, only that process is restarted.
-    #
-    # See the "Strategies" section in the `Supervisor` module for more
-    # detailed information.
-    Supervisor.start_link(children, strategy: :one_for_one, name: __MODULE__)
-  end
-end
