--- implementations/elixir/ockam/ockam_transport_udp/test/test_helper.exs
@@ -1,4 +1,0 @@
-Application.ensure_all_started(:ockam)
-Application.ensure_all_started(:ockam_transport_udp)
-
-ExUnit.start(capture_log: true, trace: true)
